requirements.txt
flask
azure-storage-blob
azure-cosmosdb-table
Flask-WTF
flask-bootstrap
Flask-Limiter
python-decouple